feat: Add API routes configuration

+++ routes/api.php
@@ -2,10 +2,14 @@
 
 use Illuminate\Support\Facades\Route;
 use Marjose123\FilamentWebhookServer\Http\Controllers\WebhookController;
+use Marjose123\FilamentWebhookServer\WebhookPlugin;
 
-Route::prefix('/webhook-server/api')->middleware('web')->group(function (): void {
-    Route::get('/', [WebhookController::class, 'get']);
-    Route::post('/', [WebhookController::class, 'create']);
-    Route::patch('/{id}', [WebhookController::class, 'update']);
-    Route::delete('/{id}', [WebhookController::class, 'delete']);
-});
+if(filament()->isServing() && WebhookPlugin::get()->isEnabled() && WebhookPlugin::get()->isApiRoutesEnabled())
+{
+    Route::prefix('/webhook-server/api')->middleware('web')->group(function (): void {
+        Route::get('/', [WebhookController::class, 'get']);
+        Route::post('/', [WebhookController::class, 'create']);
+        Route::patch('/{id}', [WebhookController::class, 'update']);
+        Route::delete('/{id}', [WebhookController::class, 'delete']);
+    });
+}
